Fixed GameConsole
Also did more work on Super Shade Bros 2, and close to done on updating the next 200 rooms, just need to playtest and fix any small bugs. Need to playtest a lot of new changes, prob move DebugConsole to Hud, add in Notes to HUD, fix achievements, and add graphics to the death screen, THEN it might be in a release state.